Refactor Gmail tagging logic and add ESLint configuration (#3)
* feat: Added eslint config + populate package json

* feat(gmail): Refactor tagEmail function to use arrow function syntax

* feat(gmail): separate logic on different files
