+++ gmail_plus_addressing_labeling/src/constants.js
@@ -1,0 +1,3 @@
+const PAGE_SIZE = 50
+const UNTAGGED = 'untagged'
+const INBOX_QUERY = 'is:inbox'
