+++ package.json
@@ -1,5 +1,35 @@
 {
+  "name": "google-scripts",
+  "version": "1.0.0",
+  "description": "A collection of useful Google Apps Scripts used to improve my Google Workspace experience",
+  "author": "Ricard Martin <32068910+rmartinbeni@users.noreply.github.com>",
+  "license": "GPL-3.0",
+  "type": "module",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/rmartinbeni/google-scripts"
+  },
+  "bugs": {
+    "url": "https://github.com/rmartinbeni/google-scripts/issues"
+  },
+  "homepage": "https://github.com/rmartinbeni/google-scripts#readme",
   "dependencies": {
     "clasp": "^1.0.0"
-  }
+  },
+  "devDependencies": {
+    "@eslint/js": "^9.9.0",
+    "eslint": "^9.9.0",
+    "eslint-plugin-googleappsscript": "^1.0.5",
+    "globals": "^15.9.0"
+  },
+  "scripts": {
+    "push-email-tag": "cd gmail_plus_addressing_labeling && clasp push",
+    "push-google-drive-manager": "cd google_drive_manager  && clasp push"
+  },
+  "keywords": [
+    "google-apps-script",
+    "google-workspace",
+    "automation",
+    "scripts"
+  ]
 }
